--- app/src/main/java/ru/practicum/android/diploma/ui/root/search/SearchScreenEvent.kt
@@ -1,6 +1,0 @@
-package ru.practicum.android.diploma.ui.root.search
-
-sealed class SearchScreenEvent {
-    data object HideKeyboard : SearchScreenEvent()
-    data object ClearSearch : SearchScreenEvent()
-}
